Extract Import into a hook
